fix: 일반 컴포넌트 -> 라우트 형식으로 변경
