+++ src/app/auctions/[auctionId]/outbid/page.tsx
@@ -1,104 +1,17 @@
 "use client";
-import React, { useState, useEffect } from 'react';
-import { 
-  ChevronLeft, 
-  Check, 
-  ChevronRight, 
-  User, 
-  Camera, 
-  Search, 
-  Home, 
-  PlusCircle, 
-  MessageCircle, 
-  Heart, 
-  Bell, 
-  Filter, 
-  Settings,
-  MoreVertical, 
-  Send, 
-  Star,
-  Clock,
-  ArrowUpRight,
-  X,
-  Trash2,
-  Eye,
-  Image as ImageIcon,
-  ArrowLeft,
-  TrendingUp,
-  Sparkles,
-  Menu,
-  ShoppingBag,
-  Store,
-  Receipt
-} from 'lucide-react';
-import { motion, AnimatePresence } from 'motion/react';
 
-import { Screen, Tab } from '../../../../types/index';
-import { ExploreIcon } from '../../../../components/common/ExploreIcon';
+import { useRouter } from "next/navigation";
 
-export default function OutbidNotificationScreen({ onBack, onProductClick, themeColor }: { onBack: () => void; onProductClick: (id: number) => void; themeColor: string; key?: string }) {
+import OutbidNotificationScreen from "./index";
+
+export default function OutbidPage() {
+  const router = useRouter();
+
   return (
-    <motion.div
-      initial={{ opacity: 0, x: 20 }}
-      animate={{ opacity: 1, x: 0 }}
-      exit={{ opacity: 0, x: -20 }}
-      className="flex-1 flex flex-col bg-white"
-    >
-      <div className="h-16 flex items-center px-4 border-b border-gray-100">
-        <button onClick={onBack} className="p-2 hover:bg-gray-100 rounded-full transition-colors">
-          <ChevronLeft size={24} />
-        </button>
-        <h1 className="flex-1 text-center font-bold text-lg mr-10">입찰 추월 알림</h1>
-      </div>
-
-      <div className="flex-1 overflow-y-auto no-scrollbar p-8 flex flex-col items-center justify-center space-y-8">
-        <div className="w-24 h-24 bg-orange-50 text-orange-500 rounded-full flex items-center justify-center">
-          <ArrowUpRight size={48} />
-        </div>
-        
-        <div className="text-center space-y-2">
-          <h2 className="text-2xl font-black text-orange-500">입찰이 추월되었습니다!</h2>
-          <p className="text-gray-500">에어팟 프로 경매에서 다른 사용자가<br />더 높은 금액을 제시했습니다.</p>
-        </div>
-
-        <div className="w-full bg-gray-50 rounded-3xl p-6 space-y-4">
-          <div className="flex items-center space-x-4">
-            <div className="w-16 h-16 rounded-xl overflow-hidden bg-gray-200">
-              <img src="https://picsum.photos/seed/p3/200/200" alt="Product" className="w-full h-full object-cover" />
-            </div>
-            <div className="flex-1">
-              <p className="text-xs text-gray-400">현재 최고가</p>
-              <p className="text-xl font-black text-red-500">₩185,000</p>
-            </div>
-          </div>
-          <div className="h-px bg-gray-200"></div>
-          <div className="space-y-2">
-            <div className="flex justify-between text-sm">
-              <span className="text-gray-400">내 입찰가</span>
-              <span className="font-bold">₩180,000</span>
-            </div>
-            <div className="flex justify-between text-sm">
-              <span className="text-gray-400">남은 시간</span>
-              <span className="font-bold text-red-500">12분 45초</span>
-            </div>
-          </div>
-        </div>
-
-        <div className="w-full space-y-3">
-          <button 
-            onClick={() => onProductClick(101)}
-            className="w-full h-16 bg-black text-white font-bold rounded-2xl text-lg shadow-lg"
-          >
-            재입찰하러 가기
-          </button>
-          <button 
-            onClick={onBack}
-            className="w-full h-16 bg-gray-100 text-gray-500 font-bold rounded-2xl"
-          >
-            포기하기
-          </button>
-        </div>
-      </div>
-    </motion.div>
+    <OutbidNotificationScreen
+      onBack={() => router.back()}
+      onProductClick={(id) => router.push(`/auctions/${id}`)}
+      themeColor="#F64257"
+    />
   );
 }
