+++ src/app/auctions/register/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from 'react';
-import RegisterScreen from '../../products/register/page';
+import RegisterScreen from '../../products/register/RegisterScreen';
 
 export default function AuctionRegisterScreen(props: any) {
   return <RegisterScreen {...props} mode="auction" />;
