+++ src/app/auctions/register/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import React from 'react';
-import RegisterScreen from '../../products/register/page';
 
-export default function AuctionRegisterScreen(props: any) {
-  return <RegisterScreen {...props} mode="auction" />;
+import RegisterScreen from "../../products/register/RegisterScreen";
+
+export default function AuctionRegisterPage() {
+  return <RegisterScreen mode="auction" />;
 }
