+++ src/app/page.tsx
@@ -56,7 +56,7 @@ import HomeScreen from "./(main)/home";
 import WishlistScreen from "./(main)/wishlist";
 import SearchScreen from "./(main)/search";
 import SearchDetailScreen from "./(main)/search/detail";
-import RegisterScreen from "./products/register/page";
+import RegisterScreen from "./products/register/RegisterScreen";
 import ChatListScreen from "./chats";
 import MyPageScreen from "./(main)/mypage";
 import ProductDetailScreen from "./products/[productId]";
